requirements.txt
yt-dlp>=2025.1.12
librosa>=0.10.2
numpy>=1.26.0
scipy>=1.12.0
soundfile>=0.12.1
pydantic>=2.8.0
requests>=2.32.0
orjson>=3.10.0
PyYAML>=6.0.2
# Optional
# essentia
# demucs
# torch
# musicbrainzngs
# google-api-python-client
